Burrows Delta: don't choke on zero-token authors
If you have any authors in your corpus who have zero tokens after tokenization then the current code gives:

    RuntimeWarning: invalid value encountered in divide

And all your values are NaN.

It wasn't obvious to me how to remove zero-token authors from the set (which would be a better solution) but we can at least make it complete successfully. This is a hack.